requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
boto3==1.34.74
botocore==1.34.74
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.1
crispy-bootstrap4==2024.1
cron-descriptor==1.4.3
cycler==0.12.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.1
django-admin-hstore-widget==1.2.1
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-ckeditor==6.7.0
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-filter==24.2
django-import-export==3.3.6
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-mptt==0.16.0
django-resized==1.0.2
django-storages==1.14.2
django-timezone-field==6.1.0
djangorestframework==3.15.1
et-xmlfile==1.1.0
flake8==7.0.0
flower==2.0.1
fonttools==4.51.0
humanize==4.9.0
idna==3.7
isort==5.13.2
jmespath==1.0.1
kiwisolver==1.4.5
kombu==5.3.5
MarkupPy==1.14
matplotlib==3.8.4
mccabe==0.7.0
numpy==1.26.4
odfpy==1.4.1
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pillow==10.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
pyparsing==3.1.2
pyTelegramBotAPI==4.18.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
requests==2.32.2
s3transfer==0.10.1
sentry-sdk==1.44.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
tornado==6.4
typing_extensions==4.11.0
tzdata==2023.4
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
xlrd==2.0.1
xlwt==1.3.0
